+++ device-management-application/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Fixed
 - Typo in conanfile url metadata
+- Warnings in device management service.
 ### Security
 
 ## [1.4.0] - 2022-7-12